driver: don't rely on a policy for access rights to the driver
Like the dbus daemon, explicitly check if a peer is privileged
when calling UpdateActivationEnvironment and BecomeMonitor. The
default policy also restricts these calls, but we should not rely
on that if we can help it.

A peer is considered to be privileged if it has UID 0 or the same
UID as the bus itself.

This makes the default policy redundant, as far as the driver itself
is concerned.

Signed-off-by: Tom Gundersen <teg@jklm.no>