dbus/socket: return SOCKET_E_SHUTDOWN on queue() if shutdown
Rather than blackhole'ing a message, return SOCKET_E_SHUTDOWN if a
socket is shutdown on queue attempts. This allows us to make the caller
decide what to do. Most importantly, the caller can now reliably select
EPOLLOUT, rather than requiring a call to socket_has_output() (which we
want to get rid of).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>