driver: validate names on request/release-name operations
The dbus-daemon reference implementation validates those names in both
cases, so we better do so as well, to stay compatible.

It would be sufficient to do that in name_new() (in src/name.c), but
that will not suffice to be error-code compatible. Hence, do it
explicitly in the handlers.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>