launch/policy-parser: introduce a PolicyParserRegistry
In follow-up commits the PolicyRegistry used in the broker will
only contain per-uid and per-gid policies, so split the
default/console/mandatory policies out as these will still be
needed for parsing, just not at runtime.

Note that at_console=true policies are now silently, dropped, we
will always assume that all clients are not on the console. This
functionality is rarely used, and has been superseeded by polkit
a long time ago.

Signed-off-by: Tom Gundersen <teg@jklm.no>