policy/peer: introduce a new PeerPolicy object
Rather than keeping one compiled policy per peer, we want to move to
keeping the policies static, and simply reference the relevant ones
from each peer.

We will need one per-UID policy, and potentially an array of per-GID
policies per peer.

For now, this just contains the compiled policy as before as the UID
policy, without any GID policies, so this is not a functional change.

Signed-off-by: Tom Gundersen <teg@jklm.no>