activation: drop path-based indexing
Instead of indexing the activation objects based on their control-path,
we leave this up to the caller to manage (which they already do).

At the same time, this now drops all back-pointers from the activation
object to the controller, so we also clean up the activation requests
to use a proper up-call to their parent interfaces.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>