name: remove comment
Drop the TODO-comment mentioning tries. We are fine with RB-Trees, and
there is no reason to move to something else *now*. If tries turn out
to be significantly faster, and the name-lookup turns out to be
crucial, lets change this later. However, it is not a security issue,
since the log(n) performance of the rb-tree is completely acceptable.
Hence, lets drop the TODO marker.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>