policy: make policies reference counted
At the moment the policy registry is guaranteed to outlive the individual
policies, and even when we gain config reload support, all the old
policies should be dropped from the peers at the same time as the policy
registry is reloaded. However, we should really not rely on this, rather
make each policy properly reference counted.

Moreover, in the future we will pin policies that will be used for
activation, and these may outlive the registry they originate from.

At the same time, do not embed the wildcard uid policy in the peer
policy, but allocate it like the other policies. This way its lifetime
is no longer tied to the peer object.

Signed-off-by: Tom Gundersen <teg@jklm.no>