activation/policy: capture the state of the sender at the time of making an activation request
The state (names and policy) of both the sender and receiver are used
to determine whether a message can be transmitted.

This is also the case for messages that cause activation. However, in this
case the situation is no longer straight-forward. Currently, we check all
policies when a message is queued on the destination, based on the states
of the sender and receiver at this time. This has two drawbacks: the sender
does not (and should not) know that the destination is being activated, and
as such it is very counter intuitive that changes they make to their state
(dropping/gaining names) after having sent a message will influence the
policy applied to the delivery of the message. More importantly, we currently
do not deliver a message at all if the peer has been disconnected by the time
the message is being queued on the destination.

We wish to change the semantisc slightly, and say that transmission policies
should be applied using the state of the sender at send and the receiver at
receive. This is unlikely to have any practical effect, as receive policies
are never (in our experience) used on method calls (which is the only thing
relevant for activation), except that messages will now (correctly) be
delivered if they originate in a peer that has been disconnected before
the destination has been activated.

For now, simply pin the state of the sender in the activation request,
and we will adjust the policy logic in follow-up commits.

Signed-off-by: Tom Gundersen <teg@jklm.no>