driver/activation: forward activation messages using state of sender at send time
In particular, the owned names and policy are pinned at SEND so any future
changes does not affect messages after they were successfully transmitted.

This makes eavesdropping and policy applicaiton more predictable from the
point of view of the sending peer. Specifically, the distinction between
regular message forwarding and name activation is reduced.

Most importantly, this means that messages are no longer dropped in case
the sending peer disconnects between the message being sent and the
name being activated.

One small quirk remains: peers eavesdropping on the unique name of the
sending peer will not receive the activation message. This is left
as a TODO, but should be figured out.

Signed-off-by: Tom Gundersen <teg@jklm.no>