test: add ./test/dbus/ directory with helpers
This adds the new ./test/dbus/ directly, together with
util-broker.[ch]. This contains a bunch of helpers to spawn our own
dbus-broker and run tests against it.

This is meant as pure unit-tests! But rather than running the tests with
the broker-code linked statically, we instead execute the broker and
verify the D-Bus behavior of it.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>