dbus/socket: allow client-side pipelining
We optimized the socket code to allow SASL+Hello+D-Bus pipelining, and
we explicitly encourage people to make use of that. However, we never
used this code for client-side D-Bus, so we never implemented
client-side pipelining.

Turns out, to make that work, there is no real code-change needed,
except reducing 'lines_done' handling to the read-side. That is, the
'lines_done' boolean now just tells us whether we consider SASL done on
the *INPUT*-side of the socket. The output side just relies on the
caller to properly switch from line to message buffers but never back.
We could add another boolean that verifies that, but the current skb
'message' pointer serves the same purpose as long as the queue is
non-empty. Hence, lets skip this extra verification and just reduce
'lines_done' to the input-side of things.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>