dbus/socket: catch MSG_CTRUNC and error out
For better debugging, catch MSG_CTRUNC and bail out. This is a fatal
error telling us that we either exceeded our own FD limit, or we
provided a too small ctrl-buffer. Both of those errors are fatal and we
better know what buffers we need. Hence, treat it as fatal error and
bail out.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>