dbus/socket: fix and document ETOOMANYREFS handling
Now that the fd-recursion fix was applied upstream, we know the
commit-sha it will end up with in Linus' tree for 4.14. So document the
status of ETOOMANYREFS in the source and explain the situation.

While at it, lets make this commit mandatory for security purposes.
Meaning, we will not add any fallback handling for ETOOMANYREFS but
rely on the recursion-limit to be dropped, just like the in-source
comment says.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>