sd-bus: make sure the bus objects are always cleaned up correctly
It is not sufficient to unref an sd_bus object, as it may contain pending
incoming or outgoing messages that pin the bus object themselves. When
releasing the last (in our case always the only) sd_bus reference held
externally we must therefore make sure that all the queues are empty.

In the normal case this means calling sd_bus_flush_close_unref() which
makes sure messages are flushed to the kernel, before any remaining ones
are dropped and the bus object unreffed (and freed, as this better be
the last reference).

In error-paths, however, we never want to block, so we skip flushing to
the kernel. Rather we accept that something has anyway gone wrong and
simply drop any pending messages. We introduce our own helper for this:
bus_close_unref(), which is exactly like sd_bus_flush_close_unref(), but
without the call to sd_bus_flush().

This is a follow-up to commit e643995264bf32c30f9c20eb0cfd36dfd7634908.

Signed-off-by: Tom Gundersen <teg@jklm.no>