dbus/socket: consider a socket to be running as long as there are pending messages
Even when the socket has been shutdown for both reading and writing, and there are
no more messages in the incomnig queue, we still need to keep keep the socket
running until the remote end has dequeued all the pending FDs. This makes sure
we never release the quota charge before the kernel no longer accounts the FDs
on the broker.

Note that if the broker exits, we will not wait for the remote end to receive
their messages, but simply release all quota charges. This means that whomever
shuts down the broker, should also be responsible for making sure all its
clients are killed to avoid any resource leakes.

Signed-off-by: Tom Gundersen <teg@jklm.no>