test/dbus/fdspam: run the spammer and the destination as separate UIDs
A client spamming itself will be disconnected, a client spamming someone
else will receive error replies in return and have its method calls dropped.

One might argue if this is the most intuitive behavior, but either way,
let's be explicit and test for the case that sender and receiver are different
users.

Signed-off-by: Tom Gundersen <teg@jklm.no>