broker: max per-UID limits configurable
Allow per-UID limits to be overriden on the commandline. It makes no sense to run a broker
without initializing the limits, nor does it make sense to change these at runtime, so we
do not expose them on the controller API, but only pass them in on the commandline.

The defaults are kept unchanged.

Signed-off-by: Tom Gundersen <teg@jklm.no>