launch/policy: import policy converter
The launcher is required to take the policy from the legacy
dbus-daemon(1) configuration files and turn it into policy suitable for
dbus-broker. We already have the pre-parsed configuration in
ConfigRoot, what we need is a converter to the suitable format.

This introduces launch/policy.[ch] which provides the converter that
imports from ConfigRoot and exports to sd_bus_message for consumption
by dbus-broker. In between it performs optimizations to reduce the
policy to a minimum.

This commit still lacks the exporter to sd_bus_message, which will be
added in a follow-up.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>