launch/policy: add export capabilities
Allow exporting the policy into an sd_bus_message. This will be used in
follow-ups to push the policy into dbus-broker.

The signature got a bit out-of-hand here. D-Bus does not seem to be
suitable to transfer structured data in bigger chunks. Meh... Lets keep
it for now and see later what we can do. It is not set in stone, yet.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>