listener: dispatch a new peer immediately after accept()
In the common case there will be a pending SASL exchange, Hello()
method call and possibly other dbus messages pending in a fd by the
time we accept() it. We want the additional latency from connecting
to the bus (compared to sending a message on an already connected peer)
to be as small as possible, so shortcut the additional pass through
the dispatch queue between accept() and the initial peer_dispatch().

Signed-off-by: Tom Gundersen <teg@jklm.no>