launch/policy: fix catching reply/error rules
The send/recv_error= and send/recv_expected_reply= attributes only
affect matching of errors and replies, but are ignored for everything
else. So lets ignore those attributes (and warn about use), but keep on
processing the rule.

But whenever we see send/recv_type=reply/error, we ignore the rule since
we know it cannot match on anything, since we never match replies.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>