launch/policy: merge default policies into each uid set
When passing the policies to dbus-broker, we have to have complete sets
for each uid. The "default" policy is the policy used if *NO* provided
uid matches. But in all other cases it will be ignored.

Hence, we have to merge all the mandatory/default/at_console rules into
each uid to make them have effect.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>