match: drop support for matching on destination
We do not error out if a destination match is installed. If it was
installed with eavesdrop=true, then that would cause an error. If
it was installed with eavesdrop=false, then it would have no effect.
If it was a monitor rule, then again destination matching has no
effect.

Signed-off-by: Tom Gundersen <teg@jklm.no>