selinux: rework helpers
Don't pretend that libselinux has no global state, but introduce
proper helpers to initialize/deinitalize it.

Make the registry refcounted, as the name associations will be
per-listener, rather than per-bus, in the same way that the
policy is.

Lastly, reduce the difference in code-paths between SELinux being
enabled and disabled. These differences are anyway not in the
fast-path, so do not optimize anything away, unless we have to.

Signed-off-by: Tom Gundersen <teg@jklm.no>