selinux: add SELinux policy files
These are not hooked up in any way, just here to provide the
documentation as to what SELinux policies should be installed in
order to be able to run the broker in enforcing mode.

Firstly the launcher binary must be labeled correctly, and secondly
it must be granted the permission to call StartTransientUnit on PID1.

Signed-off-by: Tom Gundersen <teg@jklm.no>