driver+controller: enforce NO_REPLY_EXPECTED for driver/controller calls
Never reply to method calls with NO_REPLY_EXPECTED set.

Enforce this uniformly by considering any message that should not
get a reply to have the invalid serial set, and refuse sending
replies with the invalid serial as the REPLY_SERIAL.

We are still writing out the reply messages, and discarding them,
which could be optimized to not even writing the reply.

Signed-off-by: Tom Gundersen <teg@jklm.no>