broker/seclabel: add workaround to initalize security label
SO_PEERCRED does not work as expectet for sockets created with
socketpair(2). In particular, the default label is returned, rather
than the label of the process that created the socket pair. We would
like to pin the label of the launcher as the security context for the
broker, in the same way we pin its pid and uid. However, for the time
being we will read our own (the broker's) security label from /proc,
which we set to be the same as the launcher in the SELinux policy.
This is fine for now, as long as the broker is not used in some other
context than the launcher (which we also want to support).

Signed-off-by: Tom Gundersen <teg@jklm.no>