socket: always read the full message when it is larger than the read buffer
This was always the intention.

We now ratelimit by restricting reads to IQUEUE_RECV_MAX bytes only
when we do not know how many messages are to be read. In the case
we already have the header and the pending message is larger than
the buffer, then we know that we will at most receive one message,
so there is no reason to limit the amount of data we read.

Signed-off-by: Tom Gundersen <teg@jklm.no>