util: add libaudit helpers
If libaudit is not found at compile-time, a fallback is used with
the same behavior as if audit was disabled at runtime, but without
depending on the library.

If audit is disabled, the helpers fall back to writing to stderr,
which is what the default logging of SELinux would otherwise have
done.

Signed-off-by: Tom Gundersen <teg@jklm.no>