util/selinux: use the audit subsystem for logging
We should pass all SELinux log messages to audit when available. This
is done in a similar way to how dbus-daemon(1) does it, however, we
allow memory allocation in the logging callback. This should be fine
as we consider ENOMEM to be fatal, and will exit the broker in case
that happens, unlike dbus-daemon(1) which will attempt to continue
(and hence potentially silently swallowing logging messages).

This addresses issue #18.

Signed-off-by: Tom Gundersen <teg@jklm.no>