build: drop autodetection of dependencies
We detest auto-detection during the build-process. Lets get rid of our
last workarounds for travis, and make the selection explicit.

This introduces a new option called "launcher", which defaults to true.
If set, the launcher is built.

We need to disable this for travis, since it does not provide systemd.
We will hopefully be able to switch to a proper CI soon, so this will
probably be of temporary nature, only.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>