launch: implement extended service search-paths
The search-paths for D-Bus services include more than the default
/usr/share/dbus-1/ directories. Implement the extended search-paths and
look for services in all those directories.

The only feature we don't implement is overriding XDG_DATA_DIRS via an
environemnt variable for user buses. On the system bus, this was never
supported by anyone, and on the user bus we are unaware of a use-case.
This can be added later on, if needed.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>