util: add log infrastructure
Add logging infrastructure with support for structured journal logging.

This introduces src/util/log.[ch], which provide the "struct Log"
context object to be used to stage and submit log messages to the local
logging daemon.

We support two modes:

    * stderr: In this mode only the log-message itself is streamed to
              stderr. This is the fallback mode for all legacy logging
              options.

    * journal: In this mode we submit structured log messages to a
               datagram socket provided by the caller. The format we
               follow is the systemd-journal datagram protocol. We
               allow other journal implementations just fine, though.

Internally, our log context contains a staging memfd which we map and
write to. This way, we can easily assemble bigger structured log
messages. The commit function then takes the actual log message and
submits the message to the logging daemon.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>