test/dbus: drop broken test-fdspam
The FD-Spam test always lacked a reliable target to test for. While we
can try spamming a client with FDs, we have no way to know what shall
happen to the spammer, and more importantly when it shall happen. We
could just hard-code the behavior for dbus-broker, but this is in no
way standardized.

Remove the test for now. Lets revive it once it works reliably.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>