bus/policy: index policies by uidgid ranges
For GID we only support the singleton range, but for UIDs we now allow
the policy to be indexed by a range as well.

NOTE: the range policies in the broker are stored in a datastructure
not optimized for lookup, making the algorithmic cost of connecting
peers unacceptable, at O(n log(n)) in the number of range policies.
However, our only consumer of this API (the launcher) guarantees to
only ever have at most two range policy entries. If this guarantee
no longer holds, we should revisit this and use a datastructure with
a logarithmic cost of lookup instead.

Signed-off-by: Tom Gundersen <teg@jklm.no>