build: fill NEWS with recent changes
Add recent changes to NEWS to prepare for the next release. There is
still more patches waiting to be merged before the next release, though.

Apart from recent commits, this also pseudo-bumps the libexpat
dependency to 2.2.3. Since we already require 2.2, this is only a
recommendation to update libexpat to the latest bugfix release. However,
the bugfixes affect dbus-broker-launch, so it is explicitly mentioned
here.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>