peer: pin GIDs in the peer object
The GIDs are used to compute the applied policy, so in case the policy should
be recomputed/updated it is necessary to have access to the GIDs. For debugging
purposes it could also be useful to know what GIDs apply to a given peer.

Note that if we were to simply re-query this later on we might get confused
as the fallback-path (using nss) for GID resolution does not guarantee that
the GIDs are static for the lifetime of the peer.

Signed-off-by: Tom Gundersen <teg@jklm.no>