launcher: implement SIGHUP
When SIGHUP is raised the configuration is reread from disk, and
asynchcronously (non-atomically) pushed to the broker. First all
defunct names are removed, then the policy is replaced, then new
names are added.

Note that SIGHUP is non-blocking, so a follow-up commit will
implement a blocking ReloadConfig() call on the standard
org.freedesktop.DBus interface of the broker.

Signed-off-by: Tom Gundersen <teg@jklm.no>