launcher/transient-units: use KillMode=process
This allows forked-off children to survive their parents in case the
bus-activated parent is shut down before one of their children. Moreover,
make sure to give each transient unit file a unique name, so that a
new instance can be started, even if children of an old one are still
alive.

This addresses issue #11 and #73, implementing a suggestion given in
the latter.

Reported-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>
Reported-by: Jan Alexander Steffens<jan.steffens@gmail.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>