units: make Type=notify
During start-up, the launcher will perform blocking calls (in particular
over nss) before the broker is operational. In order to avoid
dead-locks, it is important that nothing makes blocking calls over dbus
during this time (in case the launcher indirectly blocks on these other
things).

In particular, this is a problem for PID1's AddMatch calls. PID1 already
waits for dbus to become active before making any calls, however that is
not sufficient as long as dbus-broker is Type=simple and hence
considered active as soon as it is executed.

Switch to Type=notify, and indicate that dbus-broker is READY only after
the broker is fully operational, hence avoiding any deadlocks.

It is worth noting that at run-time, the launcher still can make
blocking nss calls due to ReloadConfig(), however, that is not a problem
as the broker does not block.

Reported-by: Michal Sekletár <msekleta@redhat.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>