subprojects: update
This renames c_{rbnode,list}_unlink_init() by dropping the _init()
suffix. Update all the callers to call the new function. This is not
a functional change.

Only one instance existed that already called the function without the
_init() suffix, this is left unchanged, which means that it will
reinstantiate the node after unlinking it, which it did not used to do.
However, the old behavior was likely not intentional.

Signed-off-by: Tom Gundersen <teg@jklm.no>