launch/nss: pass the correct UID with each activatable name to the broker
Resolve usernames to UIDs and pass the UID to dbus-broker for each activatable
name. This accounts resources consumed by the activatable name, before it is
activated, on the correct UID.

If no username is given, use the UID of the launcher.

Lastly, introduce an NSS cache to avoid resolving the same username repeatedly.
The cache is only kept for a single config load/reload operation, so we do not
have to worry about cache invalidation.

Signed-off-by: Tom Gundersen <teg@jklm.no>