launcher/transient: pass the UID rather than username to systemd
This should not make any practical difference, but try to be extra careful to
avoid any inconsistencies if ever systemd and dbus-broker end up parsing things
differently.

We would have liked to always pass the UID we are using for a service to systemd,
also in the case it was not explicitly configured, but systemd's user instance
does not appear to be able to deal with that, so for now at least keep setting
this property only when User was actually provided in the .service file.

Signed-off-by: Tom Gundersen <teg@jklm.no>