launcher: reduce time spent in STARTING state
dbus-broker is now in the critical path during boot, it orders before
basic.target, and as such should not block unneccessarily.

The reason for delaying at all is to make sure all blocking calls from
the launcher have finished before anything else blocks on the broker.
Move the config parsing (which does the blocking calls as part of nss)
early and immediately notify PID1 of being started as soon as that is
done.

On my machine this roughly halves the time dbus-broker blocks boot from
~60ms to ~30ms.

Signed-off-by: Tom Gundersen <teg@jklm.no>