matches: keep a MatchRegistry explicitly for NameOwnerChanged matches
Every peer and every name now has an explicit MatchRegistry for storing
NameOwnerChanges matches that may be triggered by the peer/name. Moving
matches from the driver's sender-match registry to the correct
name-owner-changed-match registry is an opt-in optimization, and is not
yet used. This change is therefore a noop.

The broadcast function in peer.c always broadcasts to subscribers to a
set of well-known names (either given explicitly, or if NULL the driver
names is assumed). Additionally, it broadcasts to a given MatchRegistry,
this used to always be the sender-match registry associated with a unique
name, but we now make it clear that there is no such assumption, and when
relevant we pass in the name-owner-changed matches here instead.

Signed-off-by: Tom Gundersen <teg@jklm.no>