launch/policy: drop eavesdropping support
The broker does not support eavesdropping and ignores the passed in 'eavesdrop' boolean.
That causes the wrong behavior in case of <deny eavesdrop=true>, which should only deny
eavesdropping, but now denies everything. Properly drop eavesdropping support from the
launcher and filter out the rules correctly.

Additionally, warn in case a rule attempts to allow eavesdropping (as the standard
session bus policy does). This is not expected to cause any problems, but the logs
should reflect the potential change of behavior.

Signed-off-by: Tom Gundersen <teg@jklm.no>