launch/policy: distinguish between primary and auxillary policies
Rather than accepting policies applying to UID-ranges, revert to
accepting policies applying to individual UID's. these are the
primary policies that apply to a UID, of which there can be at
most one per-peer.

Extend the GID policies to apply to either individual GID's or to
an UID-range. These are auxillary policies, that are ammended to
the primary one, there may be any number of auxillary policies per
peer.

The semantics is that each peer will have exactly one policy determined
by their UID, either one given explicitly, or a default one if there
is no match. Moreover, for the sake of backwards compatibility, all
matching policies corresponding to the UID-ranges and the GID's the
UID belongs to will be ammended as well. These secondary mechanisms
are discouraged.

Signed-off-by: Tom Gundersen <teg@jklm.no>