launch/policy: support explicitly specifying at_console UIDs
This is currently not hooked up, but it will allow an array of
additional UIDs to be passed in to export(), each of which should
have at_console=true policies applied (and not at_console=false
ones).

This only affects UIDs less than or equal SYSTEMUIDMAX, as all other
UIDs already get at_console=true policies.

Signed-off-by: Tom Gundersen <teg@jklm.no>