launch/config: Make correct path for included file
When config file has a include node with relative path(e.g. <include>
some.conf</include>), it makes a wrong path. The included file path
should be in the same directory as config file(denoted as prefix).

Therefore add codes to judge whether prefix is a file or directory here.
If prefix is a directory, included file's name is added after prefix.
If prefix is a file, config file's name is replaced by included file's
name. So that it can create correct path for included file.