launch/config: fix path computation for <include>
Right now, <include>some/path.conf</include> incorrectly calculates the
relative path. It appends the suffix to the path of the parent file,
rather than the parent directory.

The config_path_new() already deals with this issue. However, in case
of <include> tags we pass a forced prefix under all circumstances. This
is not what we want. This was a side effect of:

    commit dd2923eeb9fac3acce3fbf856ea6745c1624861e
    Author: Tom Gundersen <teg@jklm.no>
    Date:   Mon Aug 7 18:27:48 2017 +0200

        launch/config: support selinux_root_relative

Fix this by passing NULL. This way, config_path_new() will use the right
parent path.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>