launch/load_policy: split parsing config files from loading policy
In a follow-up patch we will use the parsed configuration also for
loading the service files.

This also allows the loading of the policy to happen after the READY
notification has been sent to systemd, as NSS is only needed for
parsing the configuration files, not for loading the policy.

Signed-off-by: Tom Gundersen <teg@jklm.no>