build: rework object-file sharing
We really want to make sure we only ever compile a source once. We do
not do per-target CFLAGS. Hence, we can safely share object code across
our targets.

For now, we kept some of the objects inside of each executable target,
though. This required recompiling those targets for our tests. This
patch changes this approach and instead introduces a global `libbus`
target, which just compiles all 'shareable' sources into a single
target and then uses it as dependency everywhere.

We rely on --gc-sections and --as-needed here. Only the code pieces
that end up being used should be linked into an executable.

While at it, we make expat only a dependency of use_launcher, rather
than a global one. This is a leftover from when the policy-parser was
still part of the broker. And this patch makes it possible to move the
expat dep into libbus.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>