bus/policy: reorder policy compilation
When compiling policies for a new user, change the order to:
    1) uid
    2) uid-ranges
    3) gids

This makes sure the most specific rules apply before less specific
rules. This is not strictly required, since all policies contain unique
priorities. However, it makes debugging policies easier, since the
behavior is consistent and follows a specific rule.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>