launch/policy: reserve default policy
The D-Bus marshaling does not support 'maybe' / 'nullable' / 'option'
types. Hence, we can never support a concept of NULL. But our policies
always require a specific connect-policy for each batch. If we have a
batch without an explicit connect-policy, we don't know which policy to
fill in. Hence, we simply use the default priority of '1' and insert a
denial-policy. We know that this is the default behavior, anyway. Hence
it cannot have any negative effect. Any connect policy that opens up the
bus must have a higher priority, hence must override the default.

However, we never correctly reserved the default policy. Hence, we might
end up with the first imported policy to get the same priority as the
default policy. Make sure to always reserve this priority index, and
the same time make sure we use a well-defined constant, rather than a
magic number.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>