match: remove all destination matching support
This never worked, and we don't think it can be made to work reliably. Instead,
simply ignore any 'destination=' argument in a match rule, and consider it a
wildcard, as if it had not been there.

This means clients will get more results than they expected, though this will
only apply to monitors, and the alternative would be to get unreliable results.

Signed-off-by: Tom Gundersen <teg@jklm.no>