launcher/nss-cache: don't hardcode the GID of the 'root' group
We hardcode the 'root' user to function even without /etc/passwd
installed and using simple policies that only distinguish between
root and non-root users. This is ok as passwd(5) documents the
root user always to have UID 0.

However, group(5) makes no such guarantee for the root group, and
also using the root group (or groups at all) is not something we
intend to optimize for nor promote.

Drop the shortcut.

Signed-off-by: Tom Gundersen <teg@jklm.no>