launcher/nss-cache: populate the cache from /etc/{passwd,group}
Before querying the cache, populate it from /etc. This short-circuits
NSS in the common case (needing to call out to NSS at all should now
likely be considered a configuration error). As NSS is fundamentally
prone to deadlocks in the interactions between its plugins and
dbus-broker, this is likely to make that less of an issue.

This does mean that the behavior we now get is as if the 'files' NSS
plugin is statically configured to take the highest precedence in
/etc/nsswitch.conf.

Signed-off-by: Tom Gundersen <teg@jklm.no>