match: refcount matches
Coalesce all identical matches into a linked list on a shared
MatchRegistryByKeys object.

In the past when sending a broadcast, we were linearly iterating all
matches on a given destination. With this (and the previous) patch(es)
we only linearly iterate all _unique_ matches on a given (destination,
object path, interface, member) tuple.

Signed-off-by: Tom Gundersen <teg@jklm.no>