driver: rework registration handling
Method calls to the empty destination are now unconditionally handled by the Peer interface,
bypassing any access restrictions.

Method calls on the driver all fail with AccessDenied if Hello() has not been called yet.
However, the connection is kept alive.

Messages for other destinations that are sent before Hello() is called on the driver cause
the offending peer to be disconnected, without returning an error to the caller.

This brings us in line with the reference implementation.

Signed-off-by: Tom Gundersen <teg@jklm.no>